pyproject.toml
[tool.poetry]
name = "repl_python3_beta_PrideNet"
version = "0.1.0"
description = ""
authors = ["neeltron <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
mysql-connector = "^2.2.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
